Refactor operator construction in py_operators
